Didn't realize the naming of wgpu_example::framework so I moved my common example utility functions into wgpu_example::utils.
