+++ examples/common/src/lib.rs
@@ -1,1 +1,2 @@
 pub mod framework;
+pub mod utils;
