+++ examples/render-to-texture/src/main.rs
@@ -1,8 +1,7 @@
 #[cfg(not(target_arch = "wasm32"))]
-use wgpu_example::framework::output_image_native;
-
+use wgpu_example::utils::output_image_native;
 #[cfg(target_arch = "wasm32")]
-use wgpu_example::framework::output_image_wasm;
+use wgpu_example::utils ::output_image_wasm;
 
 const TEXTURE_DIMS: (usize, usize) = (512, 512);
